More clean vision updates
Lots of restructuring here, before there was a significant penalty for a
player being marked as dirty, especially when they had a wide field of
vision. Instead, a player being marked as dirty just checks all the
visible coords, rather than just the intersection of dirty and visible.

The previous history on death bug has come back as a result though.