Make whiteshoe and whiteshoed not include too much
Previously they would both include the whiteshoe package, which
would load both the 'client' and 'server' modules. They now just
load the client and server respectively.

The reason for this change is that the client relies on the curses
module which is not available in pypy. The server now runs in pypy.

Signed-off-by: Jack Edge <yellowbounder@gmail.com>