Extended options for various debugging things
The -o flag provides a variety of weird and wonderful options that
probably shouldn't be used in production, but are useful for testing.
---
AlwaysDirtyPlayer will (in the event of something changing when
processing a tick or an action), recalculate all of the player's vision.

This can fix certain bugs with the vision code, and expose problems, at
the cost of making the server significantly more slow.